requirements.txt
click==7.1.1
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
Werkzeug==1.0.0
